Get the full decimal surf height from MSW

+++ surfbutler/src/datasources/MagicSeaWeedDataSource.js
@@ -48,7 +48,7 @@ const getSwellHeight = (row) => {
         .text()
         .replace('ft', '');
 
-    return parseInt(heightString);
+    return parseFloat(heightString);
 };
 
 const getSwellPeriod = (row) => {
